Documentation and small process changes
Modified Agile.md to reflect Rally better.
Modified mk-temp-dir to create a directory named ‘min’ rather than tmp.
Renamed ‘documentation’ directory to ‘docs’
Updated Release.md to follow the reality of how this repo is setup for
deployment.
Updated VC and CI to follow the reality of how this repo is setup for
test and production deployments.
Created .npmignore so tests are deployed.